requirements.txt
tensorflow==2.3.1
numpy==1.18.5
streamlit==0.69.2
Keras==2.4.3
Pillow==8.0.0
